[FEATURE] hw-10-threads: refactored. move state machine to mission processor. keep only changing of drone context in phisycs. mp decide what should be a current state and send command to physics.

+++ drone_simulation/src/simulation.cpp
@@ -129,7 +129,7 @@ void writeSimulationJSONFile(const std::vector<SimulationStep>& simSteps)
     json stepJson;
     stepJson["position"] = {{"x", step.dronePosition.x}, {"y", step.dronePosition.y}};
     stepJson["direction"] = step.droneDirection;
-    stepJson["state"] = step.droneState;
+    stepJson["state"] = static_cast<int>(step.droneState);
     stepJson["targetIndex"] = step.target;
     stepJson["dropPoint"] = {{"x", step.dropPoint.x}, {"y", step.dropPoint.y}};
     stepJson["aimPoint"] = {{"x", step.aimPoint.x}, {"y", step.aimPoint.y}};
